Refactor dev utilities into separate module
